requirements.txt
Flask
gunicorn
flask-cors
tinydb
rq
pyserial
